[Cleanup] renamed folder because of typo
