+++ lib/pages/home/plan_map/plan_itinerary_tabs/itinerary_details_expanded/leg_overview_advanced/bar_itinerary_details.dart
